// api/src/types/courses.d.ts
/*
   Copyright 2025 Champlain API Authors

   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
   You may obtain a copy of the License at

       http://www.apache.org/licenses/LICENSE-2.0

   Unless required by applicable law or agreed to in writing, software
   distributed under the License is distributed on an "AS IS" BASIS,
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.

*/

export type Course = {
    id: number,
    title: string,
    number: string,
    credit: number,
    openseats: number,
    days: string,
    times: string,
    instructor_name: string,
    description: string,
    room: string,
    subject: string,
    type: string,
    prereq: string,
    start_date: Date,
    end_date: Date,
    semesterId: number
}[];

export type semester = {
    id: number,
    name: string,
    year: string,
    date: string
};